Update AGP; add list content layout; change accent color.
<h1>Rationale:</h1>

This commit upgrades the Android Gradle Plugin to version 3.0.0-beta7 and the build tools to version 26.0.2.  It also adds the initial code for the video list and a test for it accoridingly.  This necessitated to content layouts: one for no videos entered yet (empty list) and one for 1 or more videos.

Also, the Material Design accent color has been changed to a light green.