Revisit the search activity tests to get to 100% coverage.
<h1>Rationale</h>

In this commit, the onResume() issue (whereby one line of code that is executed gets reported by Jacoco to have 2 missed branches and 5 missed instructions) is addressed.  The resolution is to eliminate the input method manager code entirely and let Android deal with it as a side effect of handling the focus.  This reasonable because there may be a hardware keyboard which makes this code moot.

Also, the list of network seeds is expanded and the code comments in SearchActivity.kt has been enhanced.

<h1>File changes:</h1>

modified:   app/src/main/java/com/pajato/argus/SearchActivity.kt

- Summary: fix onResume() to reach 100% coverage and enhance the code comments.

modified:   app/src/main/res/values/strings.xml

- Summary: enhance the network list.