Persisting and Deleting Data
# Rationale
A core component of the Argus App idea was persisting data; whenever you open up your phone, it's there. That hasn't been the case before, but it is now, and with persistent data comes the need to delete it. This has been done now rather than in a separate second commit over concerns about testing -- the tests present significantly more complex problems without the capability to delete items. As is, it's not particularly pretty to look at, but that's something we can and likely will do next week.

# Changed Files

## ListAdapter
* Handles the onClick for the new *delete* ImageButton, which deletes an entry from the adapter and the database.

## MainActivity
Now calls methods that read from the database in *onCreate* and write to it in *onActivityResult*. In addition, a separate reusable method to swap our "emptyList" or "notEmptyList" layouts. We should consider using **ListActivity** in the future, as it includes this functionality built-in.

## res/layout/video_layout
* Added an ImageButton used for deletion of items in the list.

## ActivityTestBase
* Added *before* and *after* methods that handle specifying a separate database before an activity is run and clearing up the database after the tests finish.

## ListAdapterTest
* Added new test *testPersistedData* that ensures data when added is present after lifecycle events, and data when deleted remains deleted after a lifecycle event.

# New Files

## Database.kt
* Handles all of the accessing of the SQLite Database for us, in an attempt to keep it as self-contained as possible. It contains 2 classes and an object.
* Class 1, **DatabaseEntry**, stems from an Android-Recommended approach of storing a representation of your database entries as a way to keep your database queries / writes consistent. It stores a few fields to keep things straight.
* Class 2, **DatabaseReaderHelper**, also stems from Android-Recommended approach where it does the individual accessing of the database objects.
* The Object, **DatabaseHelper**, has helper methods for writing a single video to the database, deleting a single video from the database, and getting a list of all the videos from the database. It exists to keep database calls as succinct and consistent as possible.

## res/drawable/ic_clear_black_24dp
* A simple X icon that we are currently using as the delete button icon.